+++ src/lib.rs
@@ -41,7 +41,7 @@
 //! // Send an HTTP request to the mock server. This simulates your code.
 //! let response = isahc::get(server.url("/translate?word=hello")).unwrap();
 //!
-//! // Ensure the specified mock was called exactly one time.
+//! // Ensure the specified mock was called exactly one time (or fail with a detailed error description).
 //! hello_mock.assert();
 //! // Ensure the mock server did respond as specified.
 //! assert_eq!(response.status(), 200);
@@ -1327,6 +1327,10 @@ impl Then {
     /// assert_eq!(body, "Found");
     /// assert_eq!(response.headers().get("Location").unwrap().to_str().unwrap(), "http://www.google.com");
     /// ```
+    #[deprecated(
+        since = "0.5.6",
+        note = "Please use desired response code and headers instead"
+    )]
     pub fn temporary_redirect<S: Into<String>>(mut self, redirect_url: S) -> Self {
         self.mock
             .set(self.mock.take().return_temporary_redirect(redirect_url));