+++ tests/internal/mod.rs
@@ -1,4 +1,5 @@
 mod deprecated;
+mod extensions_test;
 mod large_body_test;
 mod loop_test;
 mod runtimes_test;